Load referenced resources in HTML for web platform tests.
Before this, the plain HTML was loaded, any inline scripts were
included, and the testharness was manually inserted. This didn't run any
test files referenced from the HTML though, meaning it didn't actually
run large numbers of the tests. This fixes that, actually running the
tests for real.

In doing so, better isolation for JS code run against the window is
required, so this now uses the "vm" module in Node to do this properly.
This should be much more effective. It might reduce performance, but
this _run method shouldn't be in widespread use, so I don't think there's huge
perf concerns here.